requirements.txt
aiogram==3.20.0
openai==0.28
python-dotenv==1.0.1
